+++ v3/src/models/shared/shared-data-utils.ts
@@ -2,7 +2,6 @@ import { IAnyStateTreeNode } from "mobx-state-tree"
 import { toV2Id } from "../../utilities/codap-utils"
 import { IDataSet } from "../data/data-set"
 import { ITileContentModel } from "../tiles/tile-content"
-import { ITileModel } from "../tiles/tile-model"
 import { getSharedModelManager } from "../tiles/tile-environment"
 import {
   ISharedCaseMetadata, isSharedCaseMetadata, kSharedCaseMetadataType, SharedCaseMetadata
@@ -59,46 +58,4 @@ export function getTileCaseMetadata(tile: ITileContentModel) {
 
 export function getAllTileCaseMetadata(tile: ITileContentModel): ISharedCaseMetadata[] {
   return getTileSharedModels(tile).filter(m => isSharedCaseMetadata(m))
-}
-
-export function isTileLinkedToDataSet(tile: ITileContentModel, dataSet: IDataSet) {
-  const sharedModels = getTileSharedModels(tile)
-  return !!sharedModels.find(sharedModel => isSharedDataSet(sharedModel) && sharedModel.dataSet.id === dataSet.id)
-}
-
-export function isTileLinkedToOtherDataSet(tile: ITileContentModel, dataSet: IDataSet) {
-  const sharedModels = getTileSharedModels(tile)
-  return !!sharedModels.find(sharedModel => isSharedDataSet(sharedModel) && sharedModel.dataSet.id !== dataSet.id)
-}
-
-// removes references to the specified tile from SharedDataSet and SharedCaseMetadata tiles
-export function unlinkTileFromDataSets(tile: ITileContentModel) {
-  const sharedModelManager = getSharedModelManager(tile)
-  const sharedModels = sharedModelManager?.getTileSharedModels(tile) ?? []
-  sharedModelManager && sharedModels.forEach(sharedModel => {
-    if (sharedModel.type === kSharedDataSetType || sharedModel.type === kSharedCaseMetadataType) {
-      sharedModelManager.removeTileSharedModel(tile, sharedModel)
-    }
-  })
-}
-
-// adds references to the specified tile to the specified SharedDataSet and its SharedCaseMetadata
-export function linkTileToDataSet(tile: ITileModel, dataSet: IDataSet) {
-  const tileContent = tile.content
-  if (isTileLinkedToOtherDataSet(tileContent, dataSet)) {
-    unlinkTileFromDataSets(tileContent)
-  }
-
-  const sharedModelManager = getSharedModelManager(tile)
-  const sharedDataSets = sharedModelManager?.getSharedModelsByType<typeof SharedDataSet>(kSharedDataSetType)
-  const sharedDataSet = sharedDataSets?.find(model => model.dataSet.id === dataSet.id) as ISharedDataSet | undefined
-  if (sharedModelManager && sharedDataSet) {
-    sharedModelManager.addTileSharedModel(tileContent, sharedDataSet)
-
-    const sharedMetadata = sharedModelManager.getSharedModelsByType<typeof SharedCaseMetadata>(kSharedCaseMetadataType)
-    const sharedCaseMetadata: ISharedCaseMetadata | undefined =
-            sharedMetadata.find(model => model.data?.id === dataSet.id)
-    sharedCaseMetadata && sharedModelManager.addTileSharedModel(tileContent, sharedCaseMetadata)
-    sharedCaseMetadata?.setCaseTableTileId(tile.id)
-  }
 }
